requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
lightning-utilities==0.14.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
trajectory_distance==1.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.2.2
pillow==11.0.0
platformdirs==4.3.7
protobuf==6.30.2
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
timm==1.0.15
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchcde==0.2.5
torchdiffeq==0.2.5
torchmetrics==1.7.1
torchsde==0.2.6
torchvision==0.21.0+cu124
tqdm==4.67.1
trampoline==0.1.2
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.20.1
wheel==0.45.1
